Accessing pages

+++ www/templates/blog/blog-post.tpl
@@ -4,24 +4,32 @@
 			<div class="row">
 			
 				<div class="col-10 offset-1">
-					<div class="post-head">
+					<div class="post-head mt-30 ">
 						<h1><?=$post['title']?></h1>
-
-                        <a href="<?=HOST?>blog/post-edit?id=<?=$post['id']?>" class="button button-edit post-head--position " name="edit-button"/>Редактировать</a>
-
 					</div>
-					<div class="post-head mt-30">
-                        <a href="<?=HOST?>blog/post-delete?id=<?=$post['id']?>" class="button button-delete post-head--position post-head--position--del" name="delete-button"/>Удалить</a>
-		
-					</div>
-					<div class="post-info">
+				   <?php if( isAdmin() ) { ?>
+				
+					<div class="post-head-button ">
+                        <a href="<?=HOST?>blog/post-edit?id=<?=$post['id']?>" class=" post-head-button--edit button button-edit post-head--position " name="edit-button"/>Редактировать</a>
+	
+						<a href="<?=HOST?>blog/post-delete?id=<?=$post['id']?>" class="button button-delete post-head--position  post-head--position--del" name="delete-button"/>Удалить</a>
+				</div>
+
+				<?php } ?>
+				
+					<div class="post-info mt-30">
 						<div class="post-info__author"><?=$post['name']?> <?=$post['subname']?></div>
 						<div class="post-info__topic"><a class="postlink" href="#"><?=$post['cat_title']?></a></div>
 						<div class="post-info__date"><? echo rus_date('j F Y H:i', strtotime($post['date_time']))?></div>
 						<div class="post-info__comments"><a class="postlink" href="#">2 комментария</a></div>
 					</div>
                         <div class="post-img">
-                        <img src="<?=HOST?>usercontent/post/<?=$post['post_img']?>" alt="Горы"/>
+                       
+					    <?php if($post['post_img'] != '') { ?>
+                            <img src="<?=HOST?>usercontent/post/<?=$post['post_img']?>" alt="<?=$post['title']?>" />
+
+                        <?php } ?>
+
                     </div>
 					<div class="post-content mb-25">
 						<p class="post-content__text"><?=$post['text']?></p>
